Ahmed mahmoud

+++ AdviLaw.Domain/Entites/UserSection/User.cs
@@ -22,10 +22,13 @@ namespace AdviLaw.Domain.Entities.UserSection
         public string CountryCode { get; set; } = string.Empty;
         public string PostalCode { get; set; } = string.Empty;
 
-        public int NationalityId { get; set; }
+        public long NationalityId { get; set; }
+
         public string ImageUrl { get; set; } = string.Empty;
         public bool IsActive { get; set; }
         public Gender Gender { get; set; }
+
+
         public decimal? Balance { get; set; } = 0; // بدل الـ Account
         public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
         public DateTime UpdatedAt { get; set; } = DateTime.UtcNow;
